# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 리드마그넷 CRM API
  version: 1.0.0
  description: |
    HTML 템플릿, 캠페인, 배포 링크, 공개 신청, CRM 및 성과 API 계약입니다.
    운영자 API는 Supabase 세션 쿠키가 필요하고 공개 API는 인증 없이 사용할 수 있습니다.
servers:
  - url: /
    description: 현재 배포 환경
...
paths:
...
  /api/campaigns:
    get:
      tags: [캠페인]
      summary: 소유 캠페인 목록
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: 캠페인 목록, content: { application/json: { schema: { type: object, required: [campaigns], properties: { campaigns: { type: array, items: { $ref: '#/components/schemas/Campaign' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '500': { $ref: '#/components/responses/DatabaseError' }
        '503': { $ref: '#/components/responses/NotConfigured' }
    post:
      tags: [캠페인]
      summary: 캠페인과 공개 폼 생성
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CampaignInput' }
      responses:
        '201': { description: 생성됨, content: { application/json: { schema: { type: object, required: [campaign], properties: { campaign: { $ref: '#/components/schemas/CampaignCreated' } } } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/DatabaseError' }
        '503': { $ref: '#/components/responses/NotConfigured' }
  /api/campaigns/{id}:
    parameters: [{ $ref: '#/components/parameters/CampaignId' }]
    get:
      tags: [캠페인]
      summary: 캠페인 상세
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: 캠페인 상세, content: { application/json: { schema: { type: object, required: [campaign], properties: { campaign: { $ref: '#/components/schemas/Campaign' } } } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/DatabaseError' }
        '503': { $ref: '#/components/responses/NotConfigured' }
    patch:
      tags: [캠페인]
      summary: 캠페인 공개 상태 변경
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [published]
              properties: { published: { type: boolean } }
      responses:
        '200': { description: 공개 상태 변경됨 }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/DatabaseError' }
        '503': { $ref: '#/components/responses/NotConfigured' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: 로그인 응답이 설정하는 Supabase 세션 쿠키. 실제 쿠키 이름은 프로젝트 ref에 따라 달라집니다.
  parameters:
    CampaignId: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fieldErrors: { type: object, additionalProperties: { type: string } }
        errors: { type: array, items: { type: object, required: [code, message], properties: { code: { type: string }, message: { type: string } } } }
...
    CampaignInput:
      type: object
      required: [name, templateId, title, description, submitLabel]
      properties:
        name: { type: string, maxLength: 120 }
        templateId: { type: string, format: uuid }
        title: { type: string, maxLength: 160 }
        description: { type: string, maxLength: 1000 }
        submitLabel: { type: string, maxLength: 40 }
    CampaignForm:
      type: object
      required: [id, public_id, title, description, submit_label, template_id]
      properties:
        id: { type: string, format: uuid }
        public_id: { type: string }
        title: { type: string }
        description: { type: string }
        submit_label: { type: string }
        template_id: { type: string, format: uuid }
    Campaign:
      type: object
      required: [id, name, created_at, published_at, campaign_forms]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        created_at: { type: string, format: date-time }
        published_at: { type: [string, 'null'], format: date-time }
        campaign_forms: { type: array, items: { $ref: '#/components/schemas/CampaignForm' } }
    CampaignCreated:
      allOf:
        - $ref: '#/components/schemas/Campaign'
        - type: object
          required: [publicPath]
          properties: { publicPath: { type: string, example: /f/abc123 } }
...
  responses:
...
    ValidationError: { description: 입력값 오류, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Unauthenticated: { description: 로그인 필요, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: 소유하거나 공개된 리소스 없음, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    DatabaseError: { description: 데이터 처리 실패, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotConfigured: { description: 서버 환경변수 미설정, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...